requirements.txt
tensorflow>=2.
numpy>=1.19.2
pandas>=1.1.3
matplotlib>=3.3.2
scikit-learn>=0.23.2
seaborn>=0.11.0
streamlit>=1.2.0
pillow>=8.0.1